% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkAR}
\alias{checkAR}
\title{Check acceptance rates}
\usage{
checkAR(accept, v, verbose = TRUE)
}
\arguments{
\item{accept}{0/1 vector indicating acceptance}

\item{v}{used tuning parameter}

\item{verbose}{print progress information}
}
\value{
vector with value of tuning parameter and a boolean indicating an acceptable rate
}
\description{
The function determines whether the acceptance rate of a given vector with
acceptance states lies between 0.35 and 0.45
}
\keyword{internal}

