\docType{data}
\name{epgs}
\alias{epgs}
\title{Faecal egg count samples (before and after treatment)}
\description{
  This is an example data set containing 24 eggs per gram
  (epg) values before and after anthelmintic treatment. The
  correction factor of the diagnostic technique was 10.
}
\keyword{datasets}
