% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fc_approx_invgamma}
\alias{fc_approx_invgamma}
\title{Approximate a full conditional with given mode and curvature by an inverse gamma distribution}
\usage{
fc_approx_invgamma(mode, curvature)
}
\arguments{
\item{mode}{mode of the full conditional}

\item{curvature}{curvature at the mode}
}
\value{
shape and scale parameter of the approximating inverse gamma distribution
}
\description{
Approximate a full conditional with given mode and curvature by an inverse gamma distribution
}
\keyword{internal}

