% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MHstep}
\alias{MHstep}
\title{Do a Metropolis-Hastings step}
\usage{
MHstep(old, rprop, dprop, logUpost, ...)
}
\arguments{
\item{old}{current state of the chain}

\item{rprop}{a random number generating function with arguments n=number of random variables, ... extra-arguments}

\item{dprop}{the corresponding density function with argument x, and the same ... arguments}

\item{logUpost}{function evaluating the log unnormalized marginal posterior}

\item{\dots}{extra arguments for the proposal density and random number generator}
}
\value{
new state of the chain with a boolean indicating acceptance as attribute
}
\description{
Do a Metropolis-Hastings step
}
\keyword{internal}

