% Generated by roxygen2 (4.0.1): do not edit by hand
\name{modeCurvature_delta}
\alias{modeCurvature_delta}
\title{Compute mode and curvature at the mode of a full conditional of form
\eqn{h(x) = x^{a} (1-x)^{b} \exp(-c x)}{h(x) = x^a (1-x)^b exp(c*x)}}
\usage{
modeCurvature_delta(a, b, c)
}
\arguments{
\item{a}{a}

\item{b}{b}

\item{c}{c}
}
\value{
list with mode, curvature at the mode, and the density const*h(mode) should the mode be at the boundary
}
\description{
Compute mode and curvature at the mode of a full conditional of form
\eqn{h(x) = x^{a} (1-x)^{b} \exp(-c x)}{h(x) = x^a (1-x)^b exp(c*x)}
}
\keyword{internal}

