% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.fecrm}
\alias{print.fecrm}
\title{Print information about mcmc run}
\usage{
\method{print}{fecrm}(x, digits = 3, quantiles = c(0.025, 0.25, 0.5, 0.75,
  0.975), ...)
}
\arguments{
\item{x}{result of a call to \code{fecr_mcmc}}

\item{digits}{digits passed to printing function}

\item{quantiles}{a vector of quantiles to evaluate for each variable}
}
\value{
x
}
\description{
Print information about mcmc run
}
\author{
Michaela Paul, with contributions from Reinhard Furrer
}
\keyword{internal}

