\name{samples2mcmc}
\alias{samples2mcmc}
\title{Convert the list with samples to a \code{mcmc} object}
\usage{
samples2mcmc(x)
}
\arguments{
  \item{x}{result of a call to \code{fec_mcmc} or
  \code{fecr_mcmc}}
}
\value{
  list with \code{mcmc} objects
}
\description{
  Convert the list with samples to a \code{mcmc} object
}
