\name{setInitials1}
\alias{setInitials1}
\title{Set initial values for all parameters of the one-sample (ZI)PoGa model}
\usage{
setInitials1(initials, fec, f)
}
\arguments{
  \item{initials}{named list with initial values for
  parameters \code{mu} \code{phi} \code{mui} \code{y} and
  \code{psi}}

  \item{fec}{faecal egg count data}

  \item{f}{correction factor}
}
\value{
  A named list with initial values
}
\description{
  Set initial values for all parameters of the one-sample
  (ZI)PoGa model
}
\keyword{internal}
