% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setUpdates_PoGa_mu}
\alias{setUpdates_PoGa_mu}
\title{Specify parameter updates for PoGa (1-sample) model}
\usage{
setUpdates_PoGa_mu(priors, fec, f)
}
\arguments{
\item{priors}{list with priors for \eqn{\mu}{mu} and \eqn{\phi}{phi}}

\item{fec}{vector with faecal egg counts}

\item{f}{correction factor for the McMaster technique (e.g. 50)}
}
\value{
a named list with update functions
}
\description{
Specify parameter updates for PoGa (1-sample) model
}
\keyword{internal}

