% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setUpdates_PoGa_u}
\alias{setUpdates_PoGa_u}
\title{Specify parameter updates for PoGa_u (unpaired 2-sample) model}
\usage{
setUpdates_PoGa_u(priors, fec.pre, fec.post, f.pre, f.post)
}
\arguments{
\item{priors}{list with prior specifications}

\item{fec.pre}{vector with faecal egg counts (before treatment)}

\item{fec.post}{vector with faecal egg counts (after treatment)}

\item{f.pre}{correction factors before treatment}

\item{f.post}{correction factors after treatment}
}
\value{
a named list with update functions
}
\description{
Specify parameter updates for PoGa_u (unpaired 2-sample) model
}
\keyword{internal}

