\name{statusMsg}
\alias{statusMsg}
\title{Print information about the progress of the MCMC run}
\usage{
statusMsg(piter, time.start)
}
\arguments{
  \item{piter}{percentage of interations already done}

  \item{time.start}{starting time of the run}
}
\value{
  prints progress information about the estimated time left
}
\description{
  Print information about the progress of the MCMC run
}
\keyword{internal}
