% Generated by roxygen2 (4.0.1): do not edit by hand
\name{update_delta_beta}
\alias{update_delta_beta}
\title{Update the reduction in mean parameter \eqn{delta}
using a beta approximation to the FC as proposal}
\usage{
update_delta_beta(delta, sumYpost, sumMuPre, a, b)
}
\arguments{
\item{delta}{delta}

\item{sumYpost}{sumYpost}

\item{sumMuPre}{sumMuPre}

\item{a}{shape1 parameter of the beta prior for \eqn{delta}}

\item{b}{shape2 parameter of the beta prior for \eqn{delta}}
}
\value{
updated parameter \eqn{delta}
}
\description{
Update the reduction in mean parameter \eqn{delta}
using a beta approximation to the FC as proposal
}
\keyword{internal}

