\name{update_theta_gamma1}
\alias{update_theta_gamma1}
\title{Update a parameter \eqn{theta} based on a gamma distribution}
\usage{
update_theta_gamma1(s, r, a, b, ...)
}
\arguments{
  \item{s}{s}

  \item{r}{r}

  \item{a}{shape parameter of the gamma prior for
  \eqn{theta}}

  \item{b}{rate parameter of the gamma prior for
  \eqn{theta}}

  \item{\dots}{extra arguments (not used)}
}
\value{
  updated parameter \eqn{theta}
}
\description{
  Update a parameter \eqn{theta} based on a gamma
  distribution
}
\keyword{internal}
