% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{eggCounts-package}
\alias{eggCounts}
\alias{eggCounts-package}
\title{Hierarchical modelling of faecal egg counts}
\description{
This package implements hierarchical models for faecal egg count data to
assess anthelmintic efficacy. Bayesian inference is done via MCMC sampling.
}
\details{
\tabular{ll}{
Package: \tab eggCounts\cr
Type: \tab Package\cr
Version: \tab 0.4-1\cr
Date: \tab 2015-02-07\cr
License: \tab GPL (>= 2) \cr
LazyLoad: \tab yes\cr
}

See \code{demo("fecm", package="eggCounts")} for an illustration.
}
\author{
Michaela Paul \email{michaela.paul@uzh.ch}
}
\keyword{package}

