\docType{package}
\name{eggCounts-package}
\alias{eggCounts}
\alias{eggCounts-package}
\title{Hierarchical modelling of faecal egg counts}
\description{
This package implements Bayesian hierarchical models for the analysis of faecal egg count data. Bayesian inference is done via efficient MCMC sampling using Stan.
}
\details{
\tabular{ll}{
Package: \tab eggCounts\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2017-08-03\cr
License: \tab GPL (>= 3) \cr
LazyLoad: \tab yes\cr
}}

\author{
Craig Wang \email{craig.wang@uzh.ch} \cr
Michaela Paul 
}
\keyword{package}

