\docType{data}
\name{epgs}
\alias{epgs}
\title{Faecal egg count samples (before and after treatment)}
\usage{
data(epgs)
}
\description{
This is an example dataset containing 14 eggs per gram (epg) values in sheep
before and after anthelmintic treatment of benzimidazole.
The correction factor of the diagnostic technique was 50.
}
\format{A data.frame containing 14 observations.}
\references{
Craig Wang, Paul R. Torgerson, Johan Hoglund, Reinhard Furrer,
Zero-inflated hierarchical models for faecal egg counts to assess anthelmintic efficacy,
Veterinary Parasitology, Volume 235, 2017, Pages 20-28.
}
\keyword{datasets}

